Start of prefab support
Prefabs are UI components that on their own do not represent a finished GUI (unlike previously.) These will define the majority of elements used by GUI objects (as long as the chosen theme decides to use them) and will more generic when compared to full UI objects. Interaction happens over an interface if one is provided, or with just the whole item as a normal object if an interface/controller is not provided.

Currently WIP