Change how init functions work
Init/AfterInit functions have been replaces with RunOrder groups. All functions with the same RunOrder will run before all functions with the next RunOrder. In this way, we get the same effect as Init & AfterInit using RunOrder 1 (Init) and RunOrder 2 (AfterInit) while opening up the possibility for functions that need to run even later (such as RunOrder 3 and beyond.)

If, for example, you have a function that needs to run after everything else in all other packages, you could use a RunOrder value of 9999 to force it to run last (assuming there's not another function with the same RunOrder.)

Functions in the same module with the same RunOrder will run in the order they appear.